requirements.txt
pytest~=8.4.0
allure-pytest
selene>=2.0.b4

allure-pytest
allure-python-commons~=2.14.3